+++ integration_tests/suite/helpers/fixtures/http.py
@@ -5,6 +5,31 @@ import requests
 from functools import wraps
 
 
+def csv_source(**source_args):
+    source_args.setdefault('token', 'valid-token-master-tenant')
+    source_args.setdefault('file', '/tmp/fixture.csv')
+
+    def decorator(decorated):
+
+        @wraps(decorated)
+        def wrapper(self, *args, **kwargs):
+            client = self.get_client(source_args['token'])
+            source = client.csv_source.create(source_args)
+            try:
+                result = decorated(self, source, *args, **kwargs)
+            finally:
+                try:
+                    self.client.csv_source.delete(source['uuid'])
+                except requests.HTTPError as e:
+                    response = getattr(e, 'response', None)
+                    status_code = getattr(response, 'status_code', None)
+                    if status_code != 404:
+                        raise
+            return result
+        return wrapper
+    return decorator
+
+
 def csv_ws_source(**source_args):
     source_args.setdefault('lookup_url', 'http://example.com/fixture')
     source_args.setdefault('token', 'valid-token-master-tenant')